+++ assets/js/components/dashboard/WelcomePanel.tsx
@@ -2,12 +2,13 @@ import React from "react";
 import { getIsAuthenticated, storageGetName, storageGetPicture } from "../../utils/storageUtils";
 import UserField from "./UserField";
 import useCheckScreenType from "../../utils/hooks";
-import MembraneLogo from "./MembraneLogo";
+import MobileHeader from "./MobileHeader";
+import type { CurrentEvents } from "../../types";
 import "../../../css/dashboard/welcomepanel.css";
 
 type WelcomePanelProps = {
-  currentEvents: string;
-  setCurrentEvents: React.Dispatch<React.SetStateAction<string>>;
+  currentEvents: CurrentEvents;
+  setCurrentEvents: React.Dispatch<React.SetStateAction<CurrentEvents>>;
 };
 
 const WelcomePanel = ({ currentEvents, setCurrentEvents }: WelcomePanelProps) => {
@@ -19,26 +20,7 @@ const WelcomePanel = ({ currentEvents, setCurrentEvents }: WelcomePanelProps) =>
   return (
     <div className="WelcomePanel">
       {screenType.device == "mobile" && (
-        <>
-          <div className="LogoContainer">
-            <MembraneLogo />
-            <p className="LogoText">Membrane</p>
-          </div>
-          <div className="ModeButtonsMobile">
-            <button
-              className={`ModeButtonMobile ${currentEvents == "All events" && "Clicked"}`}
-              onClick={() => setCurrentEvents("All events")}
-            >
-              All events
-            </button>
-            <button
-              className={`ModeButtonMobile ${currentEvents == "Recorded events" && "Clicked"}`}
-              onClick={() => setCurrentEvents("Recorded events")}
-            >
-              Recorded events
-            </button>
-          </div>
-        </>
+        <MobileHeader currentEvents={currentEvents} setCurrentEvents={setCurrentEvents} />
       )}
       <div className="NamePanel">
         {currentEvents == "All events" && (
